requirements.txt
torch
torchinfo
torchvision
absl-py
certifi
charset-normalizer
easing-functions
einops
filelock
fsspec
grpcio
huggingface-hub
idna
Jinja2
llvmlite
mamba-ssm
Markdown
MarkupSafe
monai
mpmath
networkx
ninja
numba
numpy
packaging
pillow
protobuf
PyYAML
regex
requests
safetensors
setuptools
six
sympy
tensorboard
tensorboard-data-server
timm
tokenizers
tqdm
transformers
triton
typing_extensions
urllib3
Werkzeug
